refactor: change count type from float to long in stats analytics

+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/apim/core/analytics/model/StatsAnalytics.java
@@ -15,4 +15,4 @@
  */
 package io.gravitee.apim.core.analytics.model;
 
-public record StatsAnalytics(float avg, float min, float max, float sum, float count) {}
+public record StatsAnalytics(float avg, float min, float max, float sum, long count) {}
